+++ eras/alonzo/impl/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.15.0.0
 
+* Updated `AlonzoGenesis` with `extraConfig` field
+* Added `AlonzoExtraConfig` and its instances
 * Remove deprecated functions `nullRedeemers`, `lookupRedeemers`, `nullDats`
 * Remove deprecated type `AlonzoTxWits'` and its accessor functions:
   - `txwitsVKey'`
@@ -38,6 +40,7 @@
 
 ### `testlib`
 
+* Added `Arbitrary` instance for `AlonzoExtraConfig`
 * Add CDDL definitions for Plutus V1 types: `big_int`, `big_uint`, `big_nint`, `bounded_bytes`
 * Rename `plutus_script` -> `plutus_v1_script` in CDDL
 * Add `plutus_v1_script` to CDDL exports